chore: improve internal package deps
