+++ packages/cli/src/api/vite-hmr.ts
@@ -1,25 +1,6 @@
-import type { HmrUpdateEvent } from './Hmr';
+import type { SeedcordFrameworkEvents, SeedcordCliEvents } from '@seedcord/types/internal';
 
-/**
- * Map of HMR events sent from the framework (client) to the CLI (server).
- *
- * @internal
- */
-export interface SeedcordFrameworkEvents {
-    'seedcord:register-critical-files': { patterns: string[] };
-    'seedcord:commands-update-prompt': { files: string[] };
-}
-
-/**
- * Map of HMR events sent from the CLI (server) to the framework (client).
- *
- * @internal
- */
-export interface SeedcordCliEvents {
-    'seedcord:hmr': HmrUpdateEvent;
-    'seedcord:refresh-commands': { shouldRefresh: boolean };
-}
-
+// types the HMR wire events on the dev server (this.hot.send / .on) against the shared event maps.
 declare module 'vite/types/customEvent.d.ts' {
     interface CustomEventMap extends SeedcordFrameworkEvents, SeedcordCliEvents {}
 }
