+++ packages/cli/src/index.ts
@@ -6,7 +6,5 @@ export {
     type SeedcordHmrConfig
 } from '@core/config/schema';
 
-export type * from '@api/Hmr';
-
 /** Package version. Falls back to '0.0.0' when run unbuilt (dev); the real value is injected at build time. */
 export const version = process.env.PACKAGE_VERSION ?? '0.0.0';
