+++ packages/seedcord/src/hmr/HmrModuleHandler.ts
@@ -5,8 +5,8 @@ import { pathToFileURL } from 'node:url';
 import { formatFilePath } from '@seedcord/utils';
 import chalk from 'chalk';
 
-import type { HmrUpdateEvent } from '@seedcord/cli';
 import type { Logger } from '@seedcord/services';
+import type { HmrUpdateEvent } from '@seedcord/types/internal';
 
 interface HmrStore<THandler, TMiddleware, TArtifacts> {
     fileToHandlers: Map<string, Set<THandler>>;
