+++ packages/seedcord/src/vite-hmr.d.ts
@@ -1,4 +1,4 @@
-import type { SeedcordFrameworkEvents, SeedcordCliEvents } from '@seedcord/cli/vite-hmr';
+import type { SeedcordFrameworkEvents, SeedcordCliEvents } from '@seedcord/types/internal';
 
 declare module 'vite/types/customEvent.d.ts' {
     interface CustomEventMap extends SeedcordFrameworkEvents, SeedcordCliEvents {}
