+++ packages/types/src/internal.index.ts
@@ -1,1 +1,2 @@
 export * from './brand';
+export type * from './Hmr';
